Reestructured and exception handling

+++ Models/Coordinates.cs
@@ -2,7 +2,8 @@
 {
     public class Coordinates
     {
-        public float latitude { get; set; }
-        public float longitude { get; set; }
+        public double latitude { get; set; }
+        public double longitude { get; set; }
+        
     }
 }
